Implement tests which don't rely on a running MPD server
They use instead a mock-ed socket to "communicate" with the rest of the
tests which allow to control and test precisely what we want.

No tests communicate with a real MPD server anymore. If needed, we can still
get a way to add "integration" tests, or to allow some of those unittests
to communicate with a server (so-called "safe" tests).

Removed the tests which compare the commands allowed by the server and
those implemented by the library, I don't think this is the goal of the
unittest to do this, and should be done somewhere else (a separate,
dedicated script?)