unicode: lower-level encoder/decoder to deal with non-ASCII characters
On python3, socket.socket().makefile() returns a file-like object which
converts bytes to string (which is cool), using the system encoding as
default encoding/decoding method (which is not so cool).
So, if our MPD library contains non-ASCII characters and your locale is C,
_rfile.readline() will fail since it will try to decode the bytes to Unicode
using an ASCII decoder.

This patch enforces the communication with MPD to UTF-8 on Python 3, and
try to deal with it if use_unicode is True with Python 2.

Note that wrapping the file-like sockets with codecs.EncodedFile() would be
better, but it gets stuck while reading lines (incompatibility when wrapping
sockets?)