async: Tolerate early disconnects
The typing was added anticipating [145], and can be removed if that's
not going anywhere; more important was documenting the variable (even
though it's internal), and making sure it's only accessed in ways that
tolerate double disconnects (eg. due to a network disconnect racing
against a user disconnect, or because something disconnects just to be
sure).

Closes: https://github.com/Mic92/python-mpd2/issues/157

[145]: https://github.com/Mic92/python-mpd2/pull/145/files