Parse failure responses in CommandError exception
This allows for improved error handling based on error numbers, offsets
in a command list, command and ultimately also hands you the literal
error message.

The whole story is, that a home-assistant user came forward complainig
about the noise from albumart commands that returned

    [50@0] {albumart} No file exists

for every song, that didn't have any. There is currently no clean way to
ignore specific errors, but to rely on string matching. Instead of everyone
doing that for themselves I believe a simple pattern like this might be
useful.