async: Handle errors during idle
Idle errors were previously unhandled -- but can crop up when there is
authentication involved.

The idle response distribution code is refactored to follow the common
pattern of having the data preprocessing function in the callback, but
then acting on the Future result, which may be erroneous.

To avoid unnecessary API changes, errors are distributed as potential
changes to all currently observed subsystems.

Closes: https://github.com/Mic92/python-mpd2/issues/165